refactor: break circular type imports via options and grant identifier
- Move AuthorizationServerOptions interface to src/options.ts (was
  re-imported back from authorization_server.ts, creating a cycle).
- Extract GrantIdentifier type to src/grants/abstract/grant_identifier.ts
  so entity/repository/request files no longer pull in the full
  grant.interface.ts (which imports authorization.request.ts).
- Re-export both symbols from their original locations for full
  backwards compatibility — public API surface is unchanged.
- Reduces madge circular dependencies from 10 to 2 (type-only cycles
  only, no runtime impact).
