+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Internal: extract opaque-vs-JWT branching in `AuthCodeGrant` into an `AuthCodeEncoder` strategy selected once in the constructor. No public API change. JWT-mode issue and resolve continue to dispatch through the existing `protected encrypt`/`decrypt` hooks on `AbstractGrant`, so subclass overrides of those methods participate as before.
 - Internal: extract opaque-vs-JWT branching in `AbstractGrant.makeBearerTokenResponse` and `RefreshTokenGrant.validateOldRefreshToken` into a `RefreshTokenEncoder` strategy held on `AbstractGrant`. No public API change. JWT-mode issue and resolve continue to dispatch through `protected encryptRefreshToken` and `protected decrypt`, so subclass overrides of either hook participate as before.
 - Internal: drop `export` from six encoder helper types (`AuthCodeEncoderResolved`, `AuthCodeEncryptFn`, `AuthCodeDecryptFn`, `AuthCodeDecodeFn`, `RefreshTokenResolutionPayload`, `RefreshTokenResolution`) that were never reachable through the package's public surface. Emitted `.d.ts` is byte-identical, so consumers see no change.
+- Internal: break circular type imports by moving the `AuthorizationServerOptions` interface to `src/options.ts` and extracting the `GrantIdentifier` type to its own leaf file (`src/grants/abstract/grant_identifier.ts`). Both symbols are still re-exported from their original locations, so the public API is unchanged. Reduces madge-detected circular dependencies from 10 to 2.
 
 ### Fixed
 - Opaque refresh token expiration check compared seconds to a `Date`, always evaluating false and letting expired tokens through; non-expiring tokens (`refreshTokenExpiresAt: null`) were also incorrectly rejected ([#212](https://github.com/jasonraimondi/ts-oauth2-server/issues/212))
