+++ src/grants/abstract/grant_identifier.ts
@@ -1,0 +1,8 @@
+export type GrantIdentifier =
+  | "authorization_code"
+  | "client_credentials"
+  | "refresh_token"
+  | "password"
+  | "implicit"
+  | "urn:ietf:params:oauth:grant-type:token-exchange"
+  | `custom:${string}`;
