+++ src/repositories/client.repository.ts
@@ -1,5 +1,5 @@
 import { OAuthClient } from "../entities/client.entity.js";
-import { GrantIdentifier } from "../grants/abstract/grant.interface.js";
+import { GrantIdentifier } from "../grants/abstract/grant_identifier.js";
 
 export interface OAuthClientRepository {
   /**
