+++ src/repositories/scope.repository.ts
@@ -1,7 +1,7 @@
 import { OAuthClient } from "../entities/client.entity.js";
 import { OAuthScope } from "../entities/scope.entity.js";
 import { OAuthUserIdentifier } from "../entities/user.entity.js";
-import { GrantIdentifier } from "../grants/abstract/grant.interface.js";
+import { GrantIdentifier } from "../grants/abstract/grant_identifier.js";
 
 export interface OAuthScopeRepository {
   /**
