+++ src/repositories/user.repository.ts
@@ -1,6 +1,6 @@
 import { OAuthClient } from "../entities/client.entity.js";
 import { OAuthUser, OAuthUserIdentifier } from "../entities/user.entity.js";
-import { GrantIdentifier } from "../grants/abstract/grant.interface.js";
+import { GrantIdentifier } from "../grants/abstract/grant_identifier.js";
 
 export interface OAuthUserRepository {
   /**
