+++ src/requests/authorization.request.ts
@@ -3,7 +3,7 @@ import { OAuthClient } from "../entities/client.entity.js";
 import { OAuthScope } from "../entities/scope.entity.js";
 import { OAuthUser } from "../entities/user.entity.js";
 import { OAuthException } from "../exceptions/oauth.exception.js";
-import { GrantIdentifier } from "../grants/abstract/grant.interface.js";
+import { GrantIdentifier } from "../grants/abstract/grant_identifier.js";
 
 export class AuthorizationRequest {
   scopes: OAuthScope[] = [];
